// frontend/src/app/api/model/companyUpdateDto.ts
/**
 * OpenAPI definition
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {AddressDetailsDto} from './addressDetailsDto';


export interface CompanyUpdateDto {
    name: string;
    registrationNumber: string;
    addressDetails: AddressDetailsDto;
}